requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
aiofiles
jinja2
python-multipart
httpx
pytest
pytest-asyncio
alembic
psycopg2-binary
loguru
python-dotenv
passlib
bcrypt==4.0.1
pyjwt
stripe
pandas
